backend/requirements.txt
aiogram==3.23.0
fastapi==0.128.0
motor==3.6.1
PyJWT==2.9.0
pymongo
uvicorn==0.30.6
python-dotenv==1.0.1
